updated todo msg
Just pushing this to write my updated Todo because I thought of something new. We need to go through the images folder in google drive and look at all the characters we have available. If anyone has been added for which animations are currently missing, make an issue for it. Off the dome, I think this is just possibly the Delivery Driver in w0_d0 and definitely the advisor throughout the orientation. after creating issues and updating my animations proceed to todo list in previous commit